requirements.txt
torch>=1.9.0
torchvision
regex
ftfy
tqdm
pyahocorasick
